Add ELK flow collector automation to vagrant setup (#2094)

+++ test/e2e/infra/vagrant/push_antrea.sh
@@ -3,12 +3,16 @@
 function usage() {
     echo "Usage: push_antrea.sh [--prometheus] [-fc|--flow-collector <Address>] [-fa|--flow-aggregator] [-h|--help]
     Push the latest Antrea image to all vagrant nodes and restart the Antrea daemons
-          --prometheus                Deploy Prometheus service to scrape metrics from Antrea Agents and Controllers
-          --flow-collector <Address>  Provide the IPFIX flow collector address to collect the flows from the Flow Aggregator service
-                                      It should be given in the format IP:port:proto. Example: 192.168.1.100:4739:udp
-                                      Please note that with this option we deploy the Flow Aggregator Service along with Antrea.
-          --flow-aggregator           Deploy Flow Aggregator along with Antrea.
-                                      It is automatically deployed if --flow-collector is used."
+          --prometheus                 Deploy Prometheus service to scrape metrics
+                                       from Antrea Agents and Controllers.
+          --flow-collector <Addr|ELK>  Provide either the external IPFIX collector
+                                       address or specify 'ELK' to deploy the ELK
+                                       flow collector. The address should be given
+                                       in the format IP:port:proto. Example: 192.168.1.100:4739:udp.
+                                       Please note that with this option we deploy
+                                       the Flow Aggregator Service.
+          --flow-aggregator            Upload Flow Aggregator image and manifests
+                                       onto the Vagrant nodes to run Flow Aggregator e2e tests."
 }
 
 # Process execution flags
@@ -151,17 +155,33 @@ if [ "$FLOW_AGGREGATOR" == "true" ]; then
     if [[ $FLOW_COLLECTOR != "" ]]; then
         echo "Generating manifest with all features enabled along with FlowExporter feature"
         $THIS_DIR/../../../../hack/generate-manifest.sh --mode dev --all-features > "${ANTREA_YML}"
+        if [[ $FLOW_COLLECTOR == "ELK" ]]; then
+            echo "Deploy ELK flow collector"
+            echo "Copying ELK flow collector folder"
+            scp -F ssh-config -r $THIS_DIR/../../../../build/yamls/elk-flow-collector k8s-node-control-plane:~/
+            echo "Done copying"
+            # ELK flow collector needs a few minutes (2-4 mins.) to finish its deployment,
+            # so the Flow Aggregator service will not send any records till then.
+            ssh -F ssh-config k8s-node-control-plane kubectl create namespace elk-flow-collector
+            ssh -F ssh-config k8s-node-control-plane kubectl create configmap logstash-configmap -n elk-flow-collector --from-file=./elk-flow-collector/logstash/
+            ssh -F ssh-config k8s-node-control-plane kubectl apply -f elk-flow-collector/elk-flow-collector.yml -n elk-flow-collector
+            LOGSTASH_CLUSTER_IP=$(ssh -F ssh-config k8s-node-control-plane kubectl get -n elk-flow-collector svc logstash -o=jsonpath='{.spec.clusterIP}')
+            ELK_ADDR="${LOGSTASH_CLUSTER_IP}:4739:udp"
 
-        $THIS_DIR/../../../../hack/generate-manifest-flow-aggregator.sh --mode dev -fc $FLOW_COLLECTOR > "${FLOW_AGG_YML}"
+            $THIS_DIR/../../../../hack/generate-manifest-flow-aggregator.sh --mode dev -fc $ELK_ADDR > "${FLOW_AGG_YML}"
+        else
+            $THIS_DIR/../../../../hack/generate-manifest-flow-aggregator.sh --mode dev -fc $FLOW_COLLECTOR > "${FLOW_AGG_YML}"
+        fi
     else
         $THIS_DIR/../../../../hack/generate-manifest-flow-aggregator.sh --mode dev > "${FLOW_AGG_YML}"
     fi
 
     copyManifestToNodes "$FLOW_AGG_YML"
-
-    echo "Restarting Flow Aggregator deployment"
-    ssh -F ssh-config k8s-node-control-plane kubectl -n flow-aggregator delete pod --all
-    ssh -F ssh-config k8s-node-control-plane kubectl apply -f flow-aggregator.yml
+    if [[ $FLOW_COLLECTOR != "" ]]; then
+        echo "Restarting Flow Aggregator deployment"
+        ssh -F ssh-config k8s-node-control-plane kubectl -n flow-aggregator delete pod --all
+        ssh -F ssh-config k8s-node-control-plane kubectl apply -f flow-aggregator.yml
+    fi
 
     rm "${FLOW_AGG_YML}"
 fi
